--HG--
branch : yt

+++ yt/analysis_modules/level_sets/contour_finder.py
@@ -21,7 +21,7 @@ from yt.funcs import mylog, get_pbar
 from yt.utilities.lib.contour_finding import \
     ContourTree, TileContourTree, link_node_contours, \
     update_joins
-from yt.utilities.lib.grid_traversal import \
+from yt.utilities.lib.partitioned_grid import \
     PartitionedGrid
 
 def identify_contours(data_source, field, min_val, max_val,
@@ -51,7 +51,7 @@ def identify_contours(data_source, field, min_val, max_val,
             [contour_ids.view("float64")], mask,
             LE, RE, dims.astype("int64"))
         contours[nid] = (g.Level, node.node_ind, pg, sl)
-    node_ids = np.array(node_ids)
+    node_ids = np.array(node_ids).astype("int64")
     if node_ids.size == 0:
         return 0, {}
     trunk = data_source.tiles.tree.trunk
